+++ Material/Material1D/Viscosity/Kelvin.cpp
@@ -15,23 +15,23 @@
 // along with this program.  If not, see <http://www.gnu.org/licenses/>.
 ////////////////////////////////////////////////////////////////////////////////
 
-#include "Kevin.h"
+#include "Kelvin.h"
 #include <Domain/DomainBase.h>
 #include <Recorder/OutputType.h>
 
-Kevin::Kevin(const unsigned T, const unsigned DT, const unsigned ST)
+Kelvin::Kelvin(const unsigned T, const unsigned DT, const unsigned ST)
 	: Material1D(T, 0.)
 	, damper_tag(DT)
 	, spring_tag(ST) {}
 
-Kevin::Kevin(const Kevin& old_obj)
+Kelvin::Kelvin(const Kelvin& old_obj)
 	: Material1D(old_obj)
 	, damper_tag(old_obj.damper_tag)
 	, spring_tag(old_obj.spring_tag)
 	, damper(suanpan::make_copy(old_obj.damper))
 	, spring(suanpan::make_copy(old_obj.spring)) {}
 
-void Kevin::initialize(const shared_ptr<DomainBase>& D) {
+void Kelvin::initialize(const shared_ptr<DomainBase>& D) {
 	if(nullptr == D || !D->find_material(damper_tag) || !D->find_material(spring_tag)) {
 		D->disable_material(get_tag());
 		return;
@@ -51,24 +51,28 @@ void Kevin::initialize(const shared_ptr<DomainBase>& D) {
 	trial_stiffness = current_stiffness = initial_stiffness = spring->get_initial_stiffness();
 }
 
-unique_ptr<Material> Kevin::get_copy() { return make_unique<Kevin>(*this); }
+unique_ptr<Material> Kelvin::get_copy() { return make_unique<Kelvin>(*this); }
 
-int Kevin::update_trial_status(const vec& t_strain, const vec& t_strain_rate) {
+int Kelvin::update_trial_status(const vec& t_strain, const vec& t_strain_rate) {
 	incre_strain = (trial_strain = t_strain) - current_strain;
 	incre_strain_rate = (trial_strain_rate = t_strain_rate) - current_strain_rate;
 
 	if(fabs(incre_strain(0) + fabs(incre_strain_rate(0))) <= datum::eps) return SUANPAN_SUCCESS;
 
-	spring->update_trial_status(trial_strain, trial_strain_rate);
+	spring->update_trial_status(trial_strain);
 	damper->update_trial_status(trial_strain, trial_strain_rate);
 
 	trial_stiffness = spring->get_trial_stiffness();
 	trial_damping = damper->get_trial_damping();
 
+	if(!damper->get_trial_stiffness().empty()) trial_stiffness += damper->get_trial_stiffness();
+
+	trial_stress = damper->get_trial_stress() + spring->get_trial_stress();
+
 	return SUANPAN_SUCCESS;
 }
 
-int Kevin::clear_status() {
+int Kelvin::clear_status() {
 	trial_strain = current_strain.zeros();
 	trial_stress = current_stress.zeros();
 	trial_strain_rate = current_strain_rate.zeros();
@@ -77,7 +81,7 @@ int Kevin::clear_status() {
 	return spring->clear_status() + damper->clear_status();
 }
 
-int Kevin::commit_status() {
+int Kelvin::commit_status() {
 	current_strain = trial_strain;
 	current_stress = trial_stress;
 	current_strain_rate = trial_strain_rate;
@@ -86,7 +90,7 @@ int Kevin::commit_status() {
 	return spring->commit_status() + damper->commit_status();
 }
 
-int Kevin::reset_status() {
+int Kelvin::reset_status() {
 	trial_strain = current_strain;
 	trial_stress = current_stress;
 	trial_strain_rate = current_strain_rate;
@@ -95,7 +99,7 @@ int Kevin::reset_status() {
 	return spring->reset_status() + damper->reset_status();
 }
 
-vector<vec> Kevin::record(const OutputType P) {
+vector<vec> Kelvin::record(const OutputType P) {
 	vector<vec> data;
 
 	if(OutputType::SD == P || OutputType::SS == P || OutputType::S == P) data.emplace_back(current_stress);
@@ -109,4 +113,4 @@ vector<vec> Kevin::record(const OutputType P) {
 	return data;
 }
 
-void Kevin::print() { suanpan_info("A Kevin material model.\n"); }
+void Kelvin::print() { suanpan_info("A Kelvin material model.\n"); }